chore(gradle-inspector): Move two functions into the class
This is a preparation for calling a member function from these
functions.
